Add button to change background color in Form1
Implemented a new button `btnChangeBackgroundColor` in `Form1.Designer.cs` to allow users to change the form's background color to a random color. Updated the constructor in `Form1.cs` to remove the static red background and added an event handler for the button click. Additionally, improved comments and formatting in `Form1.resx` for better clarity without altering any functional resources.